Reflect, Body Size, Vip Effect, Npc Sort Enum, Gw Fix

+++ src/Comet.Game/Packets/MsgNpcInfoEx.cs
@@ -48,7 +48,7 @@ namespace Comet.Game.Packets
             PosY = npc.MapY;
             Lookface = (ushort) npc.Mesh;
             NpcType = npc.Type;
-            Sort = npc.Sort;
+            Sort = (ushort) npc.Sort;
             Name = npc.IsSynFlag() ? npc.Name : "";
         }
